requirements.txt
streamlit
anthropic
pypdf2
sentence-transformers
tiktoken
uuid # For generating unique IDs, implied by pseudo-code 
faiss-cpu
